+++ oracle-linux-image-tools/cloud/vagrant-virtualbox/image-scripts.sh
@@ -37,7 +37,8 @@ cloud::validate() {
   [[ ${VAGRANT_GUEST_ADDITIONS_URL%%:*} =~ ^((https?)|(file))$ ]] || common::error "invalid VirtualBox GA ISO URL: ${VAGRANT_GUEST_ADDITIONS_URL}"
   [[ -z "${VAGRANT_GUEST_ADDITIONS_SHA256}" ]] && common::error "missing VirtualBox GA ISO checksum"
   [[ ${#VAGRANT_GUEST_ADDITIONS_SHA256} -eq 64  ]] || common::error "VAGRANT_GUEST_ADDITIONS_SHA256 must be SHA256"
-  readonly VAGRANT_GUEST_ADDITIONS_URL VAGRANT_GUEST_ADDITIONS_SHA256
+  [[ ${VAGRANT_GUEST_ADDITIONS_KERNEL,,} =~ ^((yes)|(no))$ ]] || common::error "VAGRANT_GUEST_ADDITIONS_KERNEL must be Yes or No"
+  readonly VAGRANT_GUEST_ADDITIONS_URL VAGRANT_GUEST_ADDITIONS_SHA256 VAGRANT_GUEST_ADDITIONS_KERNEL
   # Retrieve GA during validation to "fail fast"
   declare -g VAGRANT_GUEST_ADDITIONS_PATH
   common::retrieve_iso "${VAGRANT_GUEST_ADDITIONS_URL}" "${VAGRANT_GUEST_ADDITIONS_SHA256}" VAGRANT_GUEST_ADDITIONS_PATH
@@ -95,7 +96,7 @@ cloud::sysprep_args() {
 #######################################
 cloud::image_package() {
   local mk_envelope="${CLOUD_DIR}/${CLOUD}/mk-envelope.py"
-  local cpu="${VAGRANT_VIRTUALBOX_CPU_NUM:-$CPU_NUM}"
+  local cpu="${VAGRANT_VIRTUALBOX_CPU_NUM:-${CPU_NUM%%,*}}"
   local memory="${VAGRANT_VIRTUALBOX_MEM_SIZE:-$MEM_SIZE}"
   local -a extra_disk=()
   local -a mk_envelope_params=()
@@ -144,7 +145,7 @@ cloud::image_package() {
 		load include_vagrantfile if File.exist?(include_vagrantfile)
 	EOF
 
-  if [[ "${ORACLE_RELEASE}" =~ ^[89]$ ]]; then
+  if [[ "${ORACLE_RELEASE}" =~ ^(8|9|(10))$ ]]; then
     # For the latest uek kernels (UEK7) we install kernel-uek-core which only has virtio drivers...
     mkdir "${WORKSPACE}/${VM_NAME}/include"
     cat > "${WORKSPACE}/${VM_NAME}/include/_Vagrantfile" <<-EOF