+++ oracle-linux-image-tools/distr/ol10-slim/image-scripts.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 #
-# image scripts for OL9
+# image scripts for OL10
 #
-# Copyright (c) 2022, 2025 Oracle and/or its affiliates.
+# Copyright (c) 2025 Oracle and/or its affiliates.
 # Licensed under the Universal Permissive License v 1.0 as shown at
 # https://oss.oracle.com/licenses/upl
 #